refactor(server): extract shared code from V1 API module
Move `_abs_to_rel_workspace()` from api.py to api_v2_common.py and
`create_app()` from api.py to new app.py. This separates V1 endpoints
from shared infrastructure, making future V1 removal a one-line change.

Part of Spring Cleaning #1731.
