+++ gptme/server/__init__.py
@@ -2,7 +2,7 @@
 Server for gptme.
 """
 
-from .api import create_app
+from .app import create_app
 from .cli import main
 
 __all__ = ["main", "create_app"]
