+++ gptme/server/api_v2_common.py
@@ -1,5 +1,5 @@
 """
-Common types and utilities for V2 API.
+Common types and utilities for the gptme server API.
 """
 
 from pathlib import Path
@@ -8,7 +8,22 @@ from typing import Literal, TypedDict
 from typing_extensions import NotRequired
 
 from ..message import Message
-from .api import _abs_to_rel_workspace
+from ..util.uri import URI
+
+
+def _abs_to_rel_workspace(path: str | Path | URI, workspace: Path) -> str:
+    """Convert an absolute path to a relative path.
+
+    URIs are returned as-is since they are not workspace-relative.
+    """
+    # URIs should be returned as-is (they're not workspace-relative)
+    if isinstance(path, URI):
+        return str(path)
+
+    path = Path(path).resolve()
+    if path.is_relative_to(workspace):
+        return str(path.relative_to(workspace))
+    return str(path)
 
 
 class MessageDict(TypedDict):
